PutMarkLogicRecord logging now reports the correct number of documents inserted
I have an rjrudin/42-fix-logging branch as well, but that has a lot of refactoring on it. Submitting this so that the change is isolated and easy to understand. 

Will wait to add an integration test to verify this (I think we can do it by inspecting the provenance report) until a separate is merged into develop, which has support for JUnit5-based tests.